Estimation (time series) and wavelet subtlety
Some resources about the estimation of the Hurst function and the (pointwise or
global) Hölder exponent. It will be particularly useful to help with time
series analysis.

When it comes to define the mBm, it's neceesary to take into account how the
wavelet exapnesion is used. In particular, one may define a mBm-like process that
is not a mBm per se.