First chapter revised and bibliographic addition
The first chapter's structure has been revised such that the fBm and the mBm are
introduced before harmonizable representation. The introduction has been rephrased
accordingly.

Added references from Samorodnitsky and Taqqu (1994), Nourdin (2012), and
Arcones (1995). The first one is particularly rich and useful, and may be more than
welcome to study in further details some stochastic processes.

In the preamble:
- defined "\Var" and "\Cov" commands.
- defined "presentationbox" theorem style and "presentation" theorem environment.