Proof that th fBm is not a Markov process
A proof of the non-Markov property of the fractional Brownian motion when
it is not a standard Brownian motion has been added.

Removed orthogonal decomposition (Mercer's Theorem, Karhunen-Loève decomposition).

Fixed herbin2012 ref., and added basu2007. Renamed Kolmogorov (consistency) theorem, put a placeholder for the Kolmogorov continuity theorem as well as the conditional distribution of a gaussian vector.

Smaller margins, new codename to reflect tag.